r133 Update -Fixed Exporters -OBJ, MS3D & 3DS M2 exporters now export better image names. -Added Option to use WMV M2 Model position and rotation in Exports. (Exporters need to be updated) -Flying Cameras added back in for exporting. -Patch-3.MPQ files now color green. -Lightwave Exporter Updates:  - Can now export Camera animation data.  - Creates Scenes for M2 models with camera or WMV Pos/Rot option on.

+++ trunk/src/model.h
@@ -310,7 +310,7 @@ public:
 	Vec3D *vertices, *normals;
 	Vec2D *texCoords;
 	uint16 *indices;
-	size_t nIndices;
+	uint32 nIndices;
 	std::vector<std::string> TextureList;
 	// --
